requirements.txt
Pillow
nltk
patplotlib
numpy
pandas
tweepy
textblob
python-dotenv
